refactor(client): move active-host wire facades to hostScope/activeWire (break import cycle)
recentRepos/recentAgents/savedSession/savedSessionSub/terminalListSub window
activeScope().wire, which depends on wire.ts — so defining them in wire.ts closed a
wire → hostScopes → wire import cycle that biome's noImportCycles (CI-enforced) rejects.
Move them to a leaf module hostScope/activeWire.ts that imports both wire (types) and
hostScopes (activeScope) and is imported by neither, keeping the graph acyclic. Consumers
import from the new module; their tests re-point their facade mocks accordingly.

Token: W9BE-K1RT
