+++ packages/client/src/terminal/useTerminalStore.ts
@@ -14,7 +14,7 @@ import type { TerminalId } from "kolu-common/surface";
 import { createMemo } from "solid-js";
 import { createSharedRoot } from "../createSharedRoot";
 import { useViewState } from "../useViewState";
-import { terminalListSub } from "../wire";
+import { terminalListSub } from "../hostScope/activeWire";
 import { useSubPanel } from "./useSubPanel";
 import { useTerminalMetadata } from "./useTerminalMetadata";
 import {
